Add cmake_rewrite.py and tests (#2)
This is the first in a series of PRs that will be necessary to get the updated testing system submitted. This specific change is the first of 3 substantial component rewrites. I've simplified the system used to rewrite our cmake files which had grown organically into an unmaintainable mess. This rewrite is moderately better although retains some of the structure of the original. I did not include a diff (unlike other components) since very little of the original code is retained.

The module `cmake_rewrite.py` allows the loading of a "CMakeLists.txt" file and subsequent rewriting of that file following various substitution rules including disabling of bundle lines or substitution of branches for tags.

### Other included changes.

* Add the pyproject.toml
* Add the basic skeleton of the "main.py" function used as the entrypoint.
* Add tests and test running config (using pytest)
* Some small changes to other files as necessary for `cmake_rewrite.py`.