Fixes and features for oops integration (#13)
This change was made during the oops repository integration (tested with JCSDA-internal/oops#2966). It includes both basic functionality fixes and some wish-list features that have been delayed in the old system for a long time due to the fragility and complexity of the old infrastructure making changes to the lambda a bit dangerous.

Specific changes here:

bootstrap_test.sh: uploads public logs periodically, allowing users to see build & test logs as it runs.
run_tests.sh: small bugfixes not caught in our test repo's fake builds (the test file was rewritten significantly for jedi-ci and those rewrites introduced the bug).
Test Launcher:
Cancel any old unfinished batch jobs and update the CI action with a skipped status and a "preempted" title visible in the GitHub UI.
Some exceptions are caught and raised at the end allowing the job to launch tests but notify the user of the issue.
Resolves #8