JEDI-CI performance fixes (#14)
Performance improvements for `jedi-ci`

1) __Part of the test-launcher container is pre-built and pushed to the github container repo__. This pre-building improves the speed of the action init. We do not pre-build the full image because it would prevent testing of unsubmitted changes.

2) __Use a thread pool to cancel batch jobs and GitHub check runs__. The functions used to to cancel prior tests runs are network bound and somewhat slow. Rather than running them in serial, this change runs them in parallel with a thread pool executor (Python's GIL doesn't matter here because the latency comes from non-blocking network calls).

__Why was this needed?__

The JEDI-CI launcher is somewhat interactive and a failure here causes a great deal of downstream developer pain. I want this action to execute as fast as possible in order to improve the chances that a failure is noticed before the user "walks away". Additionally, this will be JCSDA's most widely used GitHub-hosted action and this will prevent it from consuming unnecessary compute time (which is "expensive" due to our limited GitHub-earmarked funding).