Fix git credential failures (#23)
This fix appears to resolve the recently observed git credential failures during ecbuild.

Background
I'll start by noting that the following is an educate guess and a lot of the details were hard to verify since they cover the Venn diagram overlap of "poorly documented git features" and "poorly documented git integrations".

I believe that ecbuild is using git in a way that relies on simplified authentication logic; it appears to call credentials in some specific order of preference and will only use whichever credential it gets first. This is unlike typical git invocations from the shell in at least one specific case; if cached credentials are used and return an authentication error, git will attempt to fall back to a lower-preference method such as git askPass and this lower preference method will update the cache.

This issue was observed infrequently in the past an never replicated, but recently has become a very consistent cause of failure. I think this is because ecbuild has been taking a little longer lately and whereas it would not run-out the 10-minute token TTL in the past, it now does this most of the time. Because the cache was configured to last longer than the token, the cache would return an invalid token which would cause the failure then ecbuild would report the failure immediately without running any token-refresh logic.

The failure has been observed consistently and this fix worked on its first attempt. As of now my best evidence for the correctness of my theory is that this fix works. If the issue comes up again we might have to revisit.